[nnx] Object refactor
