+++ flax/experimental/nnx/__init__.py
@@ -25,7 +25,7 @@ from .nnx import errors as helpers
 from .nnx.filterlib import All as All
 from .nnx.filterlib import Not as Not
 from .nnx.graph import GraphDef as GraphDef
-from .nnx.graph import GraphNode as GraphNode
+from .nnx.object import Object as Object
 from .nnx.helpers import Dict as Dict
 from .nnx.helpers import List as List
 from .nnx.helpers import Sequential as Sequential
